Fix format errors
Ignore-this: 71bda8463c7694157628b56c780e7268

darcs-hash:ce94da2d31d01c709225057789e14de5e8288ae1

+++ source/libs/sgeobj/sge_object.c
@@ -550,7 +550,8 @@ object_get_name_prefix(const lDescr *descr, dstring *buffer)
          char *underscore = strchr(name, '_');
 
          if (underscore != NULL) {
-            sge_dstring_sprintf(buffer, "%.*s", underscore - name + 1, name);
+            sge_dstring_sprintf(buffer, "%.*s",
+                                (int) (underscore - name + 1), name);
             return sge_dstring_get_string(buffer);
          }
       }
